Module izanami::error

Error representation during handling the request.

Structs

BoxedStdCompat
A wrapper type for treating arbitrary error type as an HttpError.

Error
A type that contains arbitrary HTTP error values.

Traits

HttpError
A trait representing error values to be converted into an HTTP response.

Functions

bad_request
Equivalent to err_msg(StatusCode::BAD_REQUEST, msg).

err_msg
Creates an Error from the specified message and HTTP status code.

forbidden
Equivalent to err_msg(StatusCode::FORBIDDEN, msg).

internal_server_error
Equivalent to err_msg(StatusCode::INTERNAL_SERVER_ERROR, msg).

method_not_allowed
Equivalent to err_msg(StatusCode::METHOD_NOT_ALLOWED, msg).

not_found
Equivalent to err_msg(StatusCode::NOT_FOUND, msg).

unauthorized
Equivalent to err_msg(StatusCode::UNAUTHORIZED, msg).

Type Definitions

Result
A type alias of Result<T, E> with error::Error as error type.