Trait izanami::error::HttpError

pub trait HttpError: AsFail + Debug + Display + Send + Sync + 'static {
    fn status_code(&self) -> StatusCode;

    fn to_response(&self) -> Response<()> { ... }
}

A trait representing error values to be converted into an HTTP response.

Roughly speaking, this trait extends failure::Fail, a nearly standard of error abstraction in Rust, and provides the additional properties for handling the value as an HTTP response.

Note that this trait is defined as a sub trait of AsFail (not Fail itself), due to some restrictions of the current trait system.

Required methods

fn status_code(&self) -> StatusCode

Returns an HTTP status code associated with this error value.

Loading content...
Provided methods

fn to_response(&self) -> Response<()>

Creates an HTTP response from this error value.

Loading content...
Methods

impl dyn HttpError

pub fn is<T: HttpError>(&self) -> bool

Returns whether the concrete type of this object is the same as T or not.

pub fn downcast_ref<T: HttpError>(&self) -> Option<&T>

Attempts to downcast this object into the specified concrete type as a reference.

pub fn downcast_mut<T: HttpError>(&mut self) -> Option<&mut T>

Attempts to downcast this object into the specified concrete type as a mutable reference.

pub fn downcast<T: HttpError>(self: Box<Self>) -> Result<Box<T>, Box<Self>>

Attempts to downcast this object into the specified concrete type without unboxing.