Module izanami::rt

Miscellaneous utilities for interaction with the underlying runtime.

Re-exports

pub use tokio_threadpool::blocking as poll_blocking;

pub use tokio_threadpool::BlockingError;

Structs

BlockingSection
The future that enters a blocking section of code.

Functions

blocking_section
Creates a Future to enter the specified blocking section of code.