Function izanami::rt::blocking_section

pub fn blocking_section<F, T>(op: F) -> BlockingSection<F> where
    F: FnOnce() -> T, 

Creates a Future to enter the specified blocking section of code.

The future genereted by this function internally calls the Tokio's blocking API, and then enters a blocking section after other tasks are moved to another thread. See the documentation of tokio_threadpool::blocking for details.